% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIHKGr.R
\name{create_lihkg}
\alias{create_lihkg}
\title{Create a lihkg scraper}
\usage{
create_lihkg(..., verbose = FALSE)
}
\arguments{
\item{...}{parameters to be passed to RSelenium::rsDriver(). For example, you can use create_lihkg(browser = "firefox", port = sample(10000:60000, 1)) to generate a Selenium instance of Firefox.}

\item{verbose}{logical, whether debug information is printed out.}
}
\value{
a lihkg scraper
}
\description{
This function creates a lihkg scraper
}
